Folders are arranged. BaseRepository abstract class is added. Resource class is created for handling Api Responses
