--- android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/EventRepository.kt
@@ -1,13 +1,0 @@
-package com.example.sportsplatform.data
-
-import com.example.sportsplatform.data.models.*
-import retrofit2.Response
-
-class EventRepository(private val api: EventApi) {
-    suspend fun findEvent(eventRequest: EventRequest) : Response<EventResponse> {
-        return api.searchEvent(eventRequest.event_id)
-    }
-    suspend fun findFilterEvents(eventFilterRequest: EventFilterRequest) : Response<EventFilterResponse> {
-        return api.filterEvents(eventFilterRequest)
-    }
-}
