+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/Resource.kt
@@ -1,0 +1,12 @@
+package com.example.sportsplatform.data
+
+import okhttp3.ResponseBody
+
+sealed class Resource<out T> {
+    data class Success<out T>(val value: T) : Resource<T>()
+    data class Failure(
+        val isNetworkError: Boolean,
+        val errorCode: Int?,
+        val errorBody: ResponseBody?
+    ) : Resource<Nothing>()
+}
