--- android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/EventRequest.kt
@@ -1,5 +1,0 @@
-package com.example.sportsplatform.data.models
-
-data class EventRequest(
-    val event_id: Int
-)
