--- android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/UserFollowingResponse.kt
@@ -1,5 +1,0 @@
-package com.example.sportsplatform.data.models
-
-data class UserFollowingResponse(
-    val list_of_users : List<UserSearchResponse>
-)
