+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/requests/EventRequest.kt
@@ -1,0 +1,5 @@
+package com.example.sportsplatform.data.models.requests
+
+data class EventRequest(
+    val event_id: Int
+)
