+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/responses/TokenResponse.kt
@@ -1,4 +1,4 @@
-package com.example.sportsplatform.data.models
+package com.example.sportsplatform.data.models.responses
 
 data class TokenResponse(
     val token: String,
