+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/responses/UserFilterResponse.kt
@@ -1,4 +1,4 @@
-package com.example.sportsplatform.data.models
+package com.example.sportsplatform.data.models.responses
 
 data class UserFilterResponse(
     val total_items: Int,
