+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/models/responses/UserFollowingResponse.kt
@@ -1,0 +1,7 @@
+package com.example.sportsplatform.data.models.responses
+
+import com.example.sportsplatform.data.models.responses.UserSearchResponse
+
+data class UserFollowingResponse(
+    val list_of_users : List<UserSearchResponse>
+)
