+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/data/repository/BaseRepository.kt
@@ -1,0 +1,28 @@
+package com.example.sportsplatform.data.repository
+
+import com.example.sportsplatform.data.Resource
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import retrofit2.HttpException
+
+abstract class BaseRepository {
+
+    suspend fun <T> safeApiCall(
+        apiCall: suspend () -> T
+    ) : Resource<T> {
+        return withContext(Dispatchers.IO){
+            try {
+                Resource.Success(apiCall.invoke())
+            }catch (throwable: Throwable){
+                when(throwable){
+                    is HttpException -> {
+                        Resource.Failure(false, throwable.code(), throwable.response()?.errorBody())
+                    }
+                    else -> {
+                        Resource.Failure(true, null, null)
+                    }
+                }
+            }
+        }
+    }
+}
