+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/viewmodels/EventSearchViewModelFactory.kt
@@ -2,8 +2,7 @@ package com.example.sportsplatform.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.sportsplatform.data.EventRepository
-import com.example.sportsplatform.data.UserRepository
+import com.example.sportsplatform.data.repository.EventRepository
 
 class EventSearchViewModelFactory(
     private val repository: EventRepository
