+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/viewmodels/ProfileViewModel.kt
@@ -4,18 +4,15 @@ import android.content.Intent
 import android.view.View
 import androidx.lifecycle.ViewModel
 import com.example.sportsplatform.activities.ProfileActivity
-import com.example.sportsplatform.activities.RegisterActivity
 import com.example.sportsplatform.activities.SearchOperationsActivity
-import com.example.sportsplatform.data.EventRepository
-import com.example.sportsplatform.data.UserRepository
-import com.example.sportsplatform.data.models.EventFilterRequest
+import com.example.sportsplatform.data.repository.EventRepository
+import com.example.sportsplatform.data.repository.UserRepository
 import com.example.sportsplatform.util.Coroutines
 import com.example.sportsplatform.util.closeSoftKeyboard
-import com.example.sportsplatform.util.toast
 
 class ProfileViewModel(
-        private val userRepo: UserRepository,
-        private val eventRepo: EventRepository
+    private val userRepo: UserRepository,
+    private val eventRepo: EventRepository
 ) : ViewModel() {
 
     fun onSearchImageButtonClick(view: View) {
