+++ android/AndroidSportsPlatform/app/src/main/java/com/example/sportsplatform/viewmodels/RegisterViewModel.kt
@@ -4,9 +4,9 @@ import android.content.Intent
 import android.view.View
 import androidx.lifecycle.ViewModel
 import com.example.sportsplatform.activities.ProfileActivity
-import com.example.sportsplatform.data.UserRepository
+import com.example.sportsplatform.data.repository.UserRepository
 import com.example.sportsplatform.data.models.Sport
-import com.example.sportsplatform.data.models.UserRegisterRequest
+import com.example.sportsplatform.data.models.requests.UserRegisterRequest
 import com.example.sportsplatform.util.Coroutines
 import com.example.sportsplatform.util.toast
